+++ CHANGELOG.md
@@ -7,10 +7,11 @@
 - Modules and trial makers now accept callables for the `assets` argument, which is helpful for experiments using local file assets.
 
 ## Fixed
-- Fixed bug that was preventing `psynet simulate` from running.
+- Fixed bugs in `psynet simulate`.
 - Improved error message for `check_dallinger_version`.
 - Fixed bug in JSSynth stopAllAudio that was in some cases preventing the JSSynth from playing at all.
 - Fixed bug in `get_authenticated_session` that occasionally caused tests to fail with 'Connection reset by peer' errors.
+- Fixed bug in pgbadger workflow in CI.
 
 ## Added
 - Added adblocker note to 'Are you ready to continue?' message.